Cleanup on Proc Audio for DB work

+++ src/synesthesia/art_generator/process_audio.py
@@ -1,4 +1,8 @@
 
+from . import genre_colors, shapes_algo, curvy_algo, lines_algo, dbm
+import os
+import subprocess
+from collections import Counter
 import pickle
 import librosa
 import numpy as np
@@ -7,18 +11,32 @@ from PyQt5.QtGui import QPainterPath
 from synesthesia.wsl import *
 import warnings
 set_display_to_host()
-from collections import Counter
-import subprocess
-import os
-from . import genre_colors, shapes_algo, curvy_algo, lines_algo, dbm
 
-def proc_audio(algo,canvas, song_path, sr_selection, oct_selection, freq_scale):
+
+def drawer(canvas, algo, note, frq, oct_selection, genre):
+    if '-' in note[0]:
+        if algo == 'Shape of You':
+            shapes_algo.draw_note(canvas, note[0].split(
+                '-'), oct_selection, genre_colors.getColors(genre))
+        elif algo == 'Line Rider':
+            lines_algo.draw_note(canvas, note[0].split(
+                '-'), oct_selection, genre_colors.getColors(genre))
+        elif algo == 'Curvy':
+            path = QPainterPath()
+            curvy_algo.draw_note(canvas, note[0].split(
+                '-'), oct_selection, genre_colors.getColors(genre), path)
+
+
+def proc_audio(algo, canvas, song_path, sr_selection, oct_selection, freq_scale):
     is_mp3 = False
+    have_sample = False
     try:
-        S, bars, genre = dbm.db_driver('r', song_path, freq_scale, sr_selection)[0]
-        S = np.frombuffer(S, dtype="float32") ## Error here
+        S, bars, genre = dbm.db_driver(
+            'r', song_path, freq_scale, sr_selection)[0]
+        S = np.frombuffer(S, dtype="float32")
         bars = pickle.loads(bars)
-    except IndexError or TypeError: # when song not found
+        have_sample = True
+    except IndexError or TypeError:  # when song not found
         if song_path[-4:] == '.mp3':
             wav_path = song_path[:-4] + '.wav'
             is_mp3 = True
@@ -30,7 +48,7 @@ def proc_audio(algo,canvas, song_path, sr_selection, oct_selection, freq_scale):
             y, sr = librosa.load(song_path, sr=sr_selection)
         S = np.abs(librosa.stft(y))
         if freq_scale != 0:
-            S *= (1 +  (freq_scale / 100))
+            S *= (1 + (freq_scale / 100))
         try:
             bars = librosa.hz_to_note(S)
         except:
@@ -45,25 +63,28 @@ def proc_audio(algo,canvas, song_path, sr_selection, oct_selection, freq_scale):
             warnings.filterwarnings("default")
         except:
             genre = ''
-            
-        # TODO: store only the most common notes and freq
-        dbm.db_driver('i', song_path, freq_scale, sr_selection,S,bars,genre)
 
-    path = QPainterPath()
-    for i, bar in enumerate(bars):
-        if i == len(bars) -1:
-            pass
-        d = Counter(bar)
-        result = d.most_common(1)
-        note = result[0]
-        if '-' in note[0]:  
-            if algo == 'Shape of You':
-                shapes_algo.draw_note(canvas, note[0].split('-'), oct_selection, genre_colors.getColors(genre))
-            elif algo == 'Line Rider':
-                lines_algo.draw_note(canvas, note[0].split('-'), oct_selection, genre_colors.getColors(genre))
-            elif algo == 'Curvy':
-                curvy_algo.draw_note(canvas, note[0].split('-'), oct_selection, genre_colors.getColors(genre), path)
-            
-    if is_mp3:
-        os.remove(wav_path)
+    if not have_sample:
+        disp_notes = []
+        disp_frq = []
+        for i, bar in enumerate(bars):
+
+            c = Counter(bar)
+            result = c.most_common(1)
+            note = result[0]
+            disp_notes.append(note)
+
+            c = Counter(S[i])
+            result = c.most_common(1)
+            frq = result[0]
+            disp_frq.append(frq)
+            drawer(algo, canvas, note, frq, oct_selection, genre)
+
+        dbm.db_driver('i', song_path, freq_scale, sr_selection, S, bars, genre)
+        if is_mp3:
+            os.remove(wav_path)
+    else: # we have a sample of the audio
+        for i, note in enumerate(bars):
+            drawer(algo, canvas, note, S[i], oct_selection, genre)
+
     return True
